Allow lookup of Github OAuth token via SSM
If the `github_oauth_token` is not provided, an attempt is made to look
it up via SSM data provider. If we do explicitly pass the token into
Terraform as a variable, this value is written to the SSM param store.

The SSM key that we look for (or write) can be overridden by passing
`github_oauth_token_ssm_name` which defaults to:
`format(var.atlantis_chamber_format, var.atlantis_chamber_service, "atlantis_gh_token”`